Remove layering
